+++ mj-calc/src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Calculator from "./modules/Calculator";
+import Calculator from "./modules/Startup";
 
 function App() {
   return <Calculator />;
