+++ mj-calc/src/modules/Startup.tsx
